Create the RecordingMethodContext class.
An instance of this class is passed to LocalObject::InvokeMethod when recording execution. A separate RecordingMethodContext instance is used for each method call. This makes is easy to discard state if the method call is aborted because of a conflict.

Also, move the current transaction level to the PendingTransaction class, so that the transaction level can be easily reset if execution is rewound.