Don't use RecordingMethodContext in the implementation of PlaybackThread.
My earlier plan was that RecordingMethodContext would be the only implementation of the MethodContext interface, and that it would talk to either RecordingThread or PlaybackThread depending on whether events on the thread were being recorded or played back. However, I now think that this approach won't work because unbound objects are treated differently during playback depending on whether the thread is replaying a partial transaction or a complete transaction. For replaying part of a transaction, I need to create a new solution.